Reimplement Table in native JS
Added Native.Table (not to be confused with NaiveTable!) and Table
modules, implementing fast flat immutable arrays using native
Javascript.  This is a complete rewrite of a dependency used
extensively across the entire codebase, so much testing will be needed
to find whatever bugs it inevitably introduced.  Working with JS is
playing with fire; the type-checker won’t help us anymore.