Justify all assumeJust calls
Every call to assumeJust now has an associated justification string,
which is an informal proof for the safety of calls which the type
system cannot prove safe.  In practice, this means every indexing
operation in the codebase now has an associated comment explaining why
it can never be out of bounds, which will appear as an error message if
the logic turns out to have been incorrect.